refactor(consensus): move process_peer_msg to a separate file
